refactor(refs T30969): move DpDatepicker to components root
