+++ src/components/core/DpChangeStateAtDate.vue
@@ -65,7 +65,7 @@
 
 <script>
 import { formatDate, toDate } from '@demos-europe/demosplan-utils'
-import DpDatepicker from './form/DpDatepicker'
+import DpDatepicker from '../DpDatepicker/DpDatepicker'
 import DpLabel from '../DpLabel/DpLabel'
 import DpSelect from '../DpSelect/DpSelect'
