+++ src/components/core/form/DpDateRangePicker.vue
@@ -29,7 +29,7 @@
 </template>
 
 <script>
-import DpDatepicker from './DpDatepicker'
+import DpDatepicker from '../../DpDatepicker/DpDatepicker'
 import { formatDate } from '@demos-europe/demosplan-utils'
 
 export default {
