+++ src/components/core/form/DpDatetimePicker.vue
@@ -37,7 +37,7 @@
 <script>
 import customParseFormat from 'dayjs/plugin/customParseFormat'
 import dayjs from 'dayjs'
-import DpDatepicker from './DpDatepicker'
+import DpDatepicker from '../../DpDatepicker/DpDatepicker'
 import DpTimePicker from './DpTimePicker'
 
 dayjs.extend(customParseFormat)
